fix: Adjust spawn and door positions in YAML, change Position coordinates to double, and add nullable annotation to Door

+++ src/main/java/com/zeroninefivefive/wcnm/Arena/Stages/Position.java
@@ -6,7 +6,7 @@ import org.bukkit.World;
 import javax.annotation.Nullable;
 
 public class Position {
-    public int x, y, z;
+    public double x, y, z;
     public Location getBukkitLocation(@Nullable World World){
         return new Location(World,x,y,z);
     }
